move translate functionality to plugin translation service
